server/server/requirements.txt
-i https://pypi.org/simple
asgiref==3.7.2 ; python_version >= '3.7'
attrs==23.2.0 ; python_version >= '3.7'
autobahn==23.6.2 ; python_version >= '3.9'
automat==22.10.0
cffi==1.16.0 ; python_version >= '3.8'
channels==4.0.0
constantly==23.10.4 ; python_version >= '3.8'
cryptography==41.0.7 ; python_version >= '3.7'
daphne==4.0.0
django==4.2.1
django-admin-persian-fonts==0.2
django-better-admin-arrayfield==1.4.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-groups-manager==1.3.0
django-jet==1.0.8
django-js-asset==2.2.0 ; python_version >= '3.8'
django-jsonform==2.21.2
django-mptt==0.15.0 ; python_version >= '3.9'
django-ordered-model==3.7.4
django-simple-history==3.4.0
djangochannelsrestframework==1.2.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.1
drf-spectacular[sidecar]==0.27.0
drf-spectacular-sidecar==2024.1.1
drf-yasg==1.21.7
et-xmlfile==1.1.0 ; python_version >= '3.6'
gunicorn==21.2.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1 ; python_version >= '3.5'
jsonfield==3.1.0
jsonschema==4.20.0 ; python_version >= '3.8'
jsonschema-specifications==2023.12.1 ; python_version >= '3.8'
markdown==3.5.1
openpyxl==3.1.2
packaging==23.2 ; python_version >= '3.7'
persiantools==3.0.1
psycopg2==2.9.9
pyasn1==0.5.1 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pyasn1-modules==0.3.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
pycparser==2.21
pyjwt==2.8.0
pyopenssl==23.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
pyyaml==6.0.1 ; python_version >= '3.6'
referencing==0.32.0 ; python_version >= '3.8'
rpds-py==0.16.2 ; python_version >= '3.8'
service-identity==23.1.0
setuptools==69.0.3 ; python_version >= '3.8'
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.4 ; python_version >= '3.5'
twisted[tls]==23.10.0 ; python_full_version >= '3.8.0'
txaio==23.1.1 ; python_version >= '3.7'
typing-extensions==4.9.0 ; python_version >= '3.8'
uritemplate==4.1.1 ; python_version >= '3.6'
zope-interface==6.1 ; python_version >= '3.7'
Pillow==10.0.0
